+++ app/Filters/SoloStaffFilter.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Filters;
+
+use CodeIgniter\Filters\FilterInterface;
+use CodeIgniter\HTTP\RequestInterface;
+use CodeIgniter\HTTP\ResponseInterface;
+
+class SoloStaffFilter implements FilterInterface
+{
+    public function before(RequestInterface $request, $arguments = null)
+    {
+        $user = auth()->user();
+
+        if ($user && $user->ruolo === 'tecnico') {
+            return redirect()->to(base_url('/'))
+                ->with('error', 'Non hai i permessi per accedere a questa sezione.');
+        }
+    }
+
+    public function after(RequestInterface $request, ResponseInterface $response, $arguments = null) {}
+}